change messages to file + fix isEnableNextPage
