+++ src/client/src/components/SummaryTile/messages.ts
@@ -1,0 +1,10 @@
+
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  changeItemName: {
+    id: "summaryTile.changeItemName",
+    defaultMessage: "Change Item Name"
+  }
+});
+export default messages;
