+++ src/client/src/components/TopNavBar/messages.ts
@@ -1,0 +1,30 @@
+
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  ariaNavLabel: {
+    defaultMessage: "Navigate between pages in the Wizard",
+    id: "topNavBar.ariaNavLabel"
+  },
+  frameworks: {
+    id: "topNavBar.frameworks",
+    defaultMessage: "Add Frameworks"
+  },
+  pages: {
+    id: "topNavBar.pages",
+    defaultMessage: "Add Pages"
+  },
+  services: {
+    defaultMessage: "Add Optional Cloud Services",
+    id: "topNavBar.services"
+  },
+  summary: {
+    defaultMessage: "Summary",
+    id: "topNavBar.summary"
+  },
+  welcome: {
+    defaultMessage: "New Project",
+    id: "topNavBar.newProject"
+  }
+});
+export default messages;
