+++ src/client/src/containers/About/messages.ts
@@ -1,0 +1,26 @@
+
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  templatesVersion: {
+    id: "about.templatesVersion",
+    defaultMessage: "Templates version:"
+  },
+  wizardVersion: {
+    id: "about.wizardVersion",
+    defaultMessage: "Wizard version:"
+  },
+  reportIssue: {
+    id: "about.reportAnIssue",
+    defaultMessage: "Report an Issue"
+  },
+  visitRepo: {
+    id: "about.visitRepo",
+    defaultMessage: "Visit our GitHub"
+  },
+  feedbackRedirectLinkLabel: {
+    id: "feedback.redirectLinkLabel",
+    defaultMessage: "GitHub.com"
+  }
+});
+export default messages;
