+++ src/client/src/containers/Licenses/messages.ts
@@ -1,0 +1,10 @@
+
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  redirectLinkLabel: {
+    id: "licenses.redirectLinkLabel",
+    defaultMessage: "{licenseName} license link"
+  }
+});
+export default messages;
