+++ src/client/src/containers/SelectPages/messages.ts
@@ -1,0 +1,10 @@
+
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  pagesTitleQuestion: {
+    id: "selectPages.pagesTitleQuestion",
+    defaultMessage: "Select pages for your application"
+  }
+});
+export default messages;
