+++ src/client/src/translations/en.json
@@ -1,6 +1,4 @@
 {
-  "FunctionNameError.invalidFunctionStartLetter": "Name may only start with letters or numbers",
-  "FunctionNameError.invalidRegex": "Name may only contain letters, numbers or dashes",
   "about.reportAnIssue": "Report an Issue",
   "about.reportAnIssueLabel": "Give Feedback",
   "about.templatesVersion": "Templates version:",
@@ -101,6 +99,7 @@
   "cosmosResourceModule.createCosmosRes": "Create Cosmos DB Account",
   "cosmosResourceModule.internalName": "Internal Name",
   "cosmosResourceModule.locationSubLabel": "Choose where you want your function to be hosted",
+  "default": "default",
   "dependencyChecker.iconAltMessage": "Notification",
   "dependencyChecker.notInstalledMessage": "{dependencyName} {minimumVersion} not detected. Click to install.",
   "details.author": "Author:",
@@ -129,10 +128,6 @@
   "outputPath.ariaOutputPath": "Input for Output Path",
   "outputPath.browseButtonLabel": "Browse",
   "outputPath.outputPath": "Output Path",
-  "pageNameError.duplicateName": "Name has to be unique",
-  "pageNameError.emptyName": "Name cannot be empty",
-  "pageNameError.invalidRegex": "Name may only contain letters, numbers, spaces, dashes or underscores",
-  "pageNameError.nameStartLetter": "Page name may only start with letters",
   "pages.limitedPagesMessage": "You can select up to 20 pages",
   "pages.maxPagesText": "Notification",
   "pages.noPageGeneration": "At least 1 page must be selected",
@@ -162,7 +157,6 @@
   "projectName.ariaProjectName": "Project Name Input",
   "projectName.outputPathTitle": "Output Path:",
   "projectName.projectNameTitle": "Project Name:",
-  "projectNameError.invalidRegex": "Name may only contain letters, numbers, dashes or underscores",
   "quickStart.button": "Quick Start Web Project",
   "quickStart.description": "To get started quickly with your preferred frameworks and just a blank page, use quick start.",
   "quickStart.optional": "OPTIONAL",
@@ -186,6 +180,9 @@
   "selectFrontendFramework.selectFrontendFramework": "Select a front-end framework",
   "selectPages.pagesTitleQuestion": "Select pages for your application",
   "selectPages.webAppTitleQuestion": "Select a project type",
+  "selectableCard.details": "Learn more",
+  "selectableCard.pageCount": "{number} {page}",
+  "selectableCard.preview": "Preview",
   "sortablePageList.duplicateName": "page name has to be unique",
   "sortablePageList.hide": "Hide",
   "sortablePageList.pages": "Pages",
@@ -199,5 +196,10 @@
   "topNavBar.pages": "Pages",
   "topNavBar.services": "Add Optional Cloud Services",
   "topNavBar.summary": "Summary",
+  "validations.comun.duplicateName": "Name is reserved",
+  "validations.comun.emptyName": "Name cannot be empty",
+  "validations.comun.nameStartLetter": "Name may only start with letters",
+  "validations.comun.nameStartWith$": "Name cannot begin with $",
+  "validations.name.nameContainLettersNumbersDashes": "Name may only contain letters, numbers or dashes",
   "viewLicensesModal.closeModalLabel": "Close"
 }
